+++ tokenized_midi_datasets/AwesomeTokensDataset/AwesomeTokensDataset.py
@@ -1,12 +1,11 @@
-import fortepyan as ff
+from datasets import DatasetInfo
 from hydra.utils import to_absolute_path
-from datasets import Dataset, DatasetInfo
 from midi_trainable_tokenizers.awesome_midi_tokenzier import AwesomeMidiTokenizer
 
 from tokenized_midi_datasets import TokenizedMidiDataset
 
 _DESC = """
-Dataset with midi files, tokenzied using OneTimeTokenizer, with records of equal size.
+Dataset with midi files, tokenzied using AwesomeTokensTokenizer, with records of equal size.
 """
 
 
@@ -14,17 +13,7 @@ class AwesomeTokensDataset(TokenizedMidiDataset):
     def _info(self) -> DatasetInfo:
         return DatasetInfo(description=_DESC)
 
-    def _generate_examples(self, dataset_shards: list[Dataset]):
+    def load_tokenizer(self):
         pretrained_path = to_absolute_path("pretrained/awesome_tokenizers/awesome-tokenizer-pretrained.json")
-        self.tokenizer = AwesomeMidiTokenizer.from_file(pretrained_path)
-
-        for shard_id, dataset in enumerate(dataset_shards):
-            for it, record in enumerate(dataset):
-                piece = ff.MidiPiece.from_huggingface(dict(record))
-
-                pieces = self.filter_pauses(piece)
-                chopped_sequences = sum([self.piece_to_records(piece) for piece in pieces], [])
-
-                for jt, sequence in enumerate(chopped_sequences):
-                    key = f"{it}_{jt}_{shard_id}"  # for some reason there was duplicate key error without shard_id
-                    yield key, sequence
+        tokenizer = AwesomeMidiTokenizer.from_file(pretrained_path)
+        return tokenizer
